separate exporters into a http port for each instance
This modification removes multiprocessing and replaces it with multithreading
Also, Flask is removed from the project and replaced with http.server library
We now rely on http_sd_config functionality from Prometheus and start a
http server for each instance. This reduces the amount of data sent when
all the instances were probed through the single port.